refactor: move useswipeinteraction hook to shared homeview (#1598)
