+++ app/components/Home/components/Section/components/ga2/SectionAnalyticsAndData/components/AnalyticsToolsAndData/hooks/useInteractiveAnalyticsAndData/common/entities.ts
@@ -1,5 +1,5 @@
-import { type UseSwipeInteraction } from "@/hooks/useSwipeInteraction/useSwipeInteraction";
 import { type CardProps as DXCardProps } from "@databiosphere/findable-ui/lib/components/common/Card/card";
+import { type UseSwipeInteraction } from "@repo/shared/views/HomeView/hooks/UseSwipeInteraction/types";
 
 export interface UseInteractiveAnalyticsAndData {
   activeIndex: UseSwipeInteraction["activeIndex"];
