+++ packages/shared/views/HomeView/hooks/UseSwipeInteraction/constants.ts
